requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
opencv-contrib-python-headless
numpy
pandas
matplotlib
pillow
pyyaml
requests
scipy
scikit-learn
torch
torchvision
torchaudio
mediapipe
imageio
imageio-ffmpeg
SpeechRecognition
pyaudio
pyttsx3
gTTS
playsound
